챌린지 3회 참여제한 (#57)

+++ src/main/java/com/knj/mirou/boundedContext/challengemember/repository/ChallengeMemberRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 public interface ChallengeMemberRepository extends JpaRepository<ChallengeMember, Long> {
 
-    long countByLinkedMemberAndProgress(Member member, Progress progress);
+    int countByLinkedMemberAndProgress(Member member, Progress progress);
 
     Optional<ChallengeMember> findByLinkedMember(Member linkedMember);
